+++ __tests__/changelog-updater.test.ts
@@ -23,7 +23,7 @@ const CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES = `# Changelog
 test('adds an entry to the changelog when section already exists with entry', async () => {
   mockReadStream(CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -44,7 +44,7 @@ const CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES_UNRELEASED = `# Changelog
 test('adds an entry to the changelog when section exists under default unreleased version', async () => {
   mockReadStream(CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES_UNRELEASED)
 
-  await updateChangelog(PACKAGE_ENTRY, 'nope', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'nope', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -64,7 +64,7 @@ const CHANGELOG_WITH_PROPER_SECTIONS_UNRELEASED = `# Changelog
 test('adds an entry to the changelog when section already exists, but no entry', async () => {
   mockReadStream(CHANGELOG_WITH_PROPER_SECTIONS_UNRELEASED)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -84,7 +84,7 @@ test('adds an entry to the changelog when version exists but section does not', 
   fs.createReadStream.mockReturnValue(readable)
   fs.readFileSync.mockReturnValue(CHANGELOG_MISSING_DEPENDECIES)
 
-  await updateChangelog(PACKAGE_ENTRY, 'UNRELEASED', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'UNRELEASED', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -109,7 +109,7 @@ const CHANGELOG_WITH_MULTIPLE_VERSIONS = `# Changelog
 test('adds an entry to the changelog - multiple versions', async () => {
   mockReadStream(CHANGELOG_WITH_MULTIPLE_VERSIONS)
 
-  await updateChangelog(PACKAGE_ENTRY, 'UNRELEASED', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'UNRELEASED', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -132,7 +132,7 @@ test('errors when there is no version section', async () => {
   mockReadStream(CHANGELOG_WITH_NO_VERSION)
 
   try {
-    await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+    await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
   } catch (err) {
     expect(err).not.toBeNull()
     expect(fs.writeFileSync).toBeCalledTimes(0)
@@ -148,7 +148,7 @@ const CHANGELOG_WITH_DUPLICATE_ENTRY = `# Changelog
 test('does not update the changelog on duplicate entry', async () => {
   mockReadStream(CHANGELOG_WITH_DUPLICATE_ENTRY)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   expect(fs.writeFileSync).toBeCalledTimes(0)
 })
@@ -166,7 +166,7 @@ const CHANGELOG_WITH_DUPLICATE_ENTRY_NOT_LAST_LINE = `# Changelog
 test('does not update the changelog on duplicate entry when not the last item', async () => {
   mockReadStream(CHANGELOG_WITH_DUPLICATE_ENTRY_NOT_LAST_LINE)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   expect(fs.writeFileSync).toBeCalledTimes(0)
 })
@@ -180,7 +180,7 @@ const CHANGELOG_WITH_ENTRY_TO_UPDATE = `# Changelog
 test('updates an entry for an existing package in the same version', async () => {
   mockReadStream(CHANGELOG_WITH_ENTRY_TO_UPDATE)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -202,7 +202,7 @@ const CHANGELOG_WITH_VERSION_MISSING_DEP_SECTION_BUT_HAS_OTHERS = `# Changelog
 test('updates version with new section and entry', async () => {
   mockReadStream(CHANGELOG_WITH_VERSION_MISSING_DEP_SECTION_BUT_HAS_OTHERS)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -230,7 +230,7 @@ const CHANGELOG_WITH_MULTI_VERSION_PACKAGE_UPDATES = `# Changelog
 test('Does not update lines additional times', async () => {
   mockReadStream(CHANGELOG_WITH_MULTI_VERSION_PACKAGE_UPDATES)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -265,7 +265,7 @@ const CHANGELOG_WITH_EXISTING_SECTION_AND_SEPARATED_SECTIONS = `# Changelog
 test('Updates existing section when sections separated by blank lines', async () => {
   mockReadStream(CHANGELOG_WITH_EXISTING_SECTION_AND_SEPARATED_SECTIONS)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -306,7 +306,7 @@ const CHANGELOG_WITHOUT_EXISTING_SECTION_AND_SEPARATED_SECTIONS = `# Changelog
 test('Adds section when sections separated by blank lines', async () => {
   mockReadStream(CHANGELOG_WITHOUT_EXISTING_SECTION_AND_SEPARATED_SECTIONS)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -351,7 +351,7 @@ const CHANGELOG_WITH_EXISTING_SECTION_BETWEEN_OTHERS = `# Changelog
 test('Updates existing section when between other sections', async () => {
   mockReadStream(CHANGELOG_WITH_EXISTING_SECTION_BETWEEN_OTHERS)
 
-  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md')
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Bump')
 
   const params = fs.writeFileSync.mock.calls[0]
   expect(params[0]).toStrictEqual('./CHANGELOG.md')
@@ -375,6 +375,67 @@ test('Updates existing section when between other sections', async () => {
 ### Dependencies
 - Bump \`package\` from alpha to v1`
   )
+})
+
+const CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES_DIFFERENT_PREFIX = `# Changelog
+
+## [v1.0.0]
+### Dependencies
+- Bump \`different-package\` from v1 to v2`
+
+test('adds an entry with a different prefix to the changelog when section already exists with entry', async () => {
+  mockReadStream(CHANGELOG_WITH_PROPER_SECTIONS_AND_ENTRIES)
+
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Update')
+
+  const params = fs.writeFileSync.mock.calls[0]
+  expect(params[0]).toStrictEqual('./CHANGELOG.md')
+  expect(params[1]).toStrictEqual(`# Changelog
+
+## [v1.0.0]
+### Dependencies
+- Bump \`different-package\` from v1 to v2
+- Update \`package\` from v1 to v2`)
+})
+
+const CHANGELOG_WITH_DUPLICATE_ENTRY_DIFFERENT_PREFIX = `# Changelog
+
+## [v1.0.0]
+### Dependencies
+- Bump \`package\` from v1 to v2`
+
+test('keeps prefix on entry with a different prefix but is otherwise a duplicate', async () => {
+  mockReadStream(CHANGELOG_WITH_DUPLICATE_ENTRY_DIFFERENT_PREFIX)
+
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Update')
+
+  const params = fs.writeFileSync.mock.calls[0]
+  expect(params[0]).toStrictEqual('./CHANGELOG.md')
+  expect(params[1]).toStrictEqual(`# Changelog
+
+## [v1.0.0]
+### Dependencies
+- Bump \`package\` from v1 to v2`)
+})
+
+const CHANGELOG_WITH_EXISTING_ENTRY_DIFFERENT_PREFIX = `# Changelog
+
+## [v1.0.0]
+### Dependencies
+- Bump \`package\` from v1 to v1.1`
+
+test('keeps prefix on entry with a different prefix', async () => {
+  mockReadStream(CHANGELOG_WITH_DUPLICATE_ENTRY_DIFFERENT_PREFIX)
+
+  await updateChangelog(PACKAGE_ENTRY, 'v1.0.0', './CHANGELOG.md', 'Update')
+
+  const params = fs.writeFileSync.mock.calls[0]
+  expect(params[0]).toStrictEqual('./CHANGELOG.md')
+  expect(params[1]).toStrictEqual(`# Changelog
+
+## [v1.0.0]
+### Dependencies
+- Bump \`package\` from v1 to v2`)
 })
 
 function mockReadStream(changelog: string) {