+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [UNRELEASED]
+### Changed
+- Use imperative mood in changelog entry by default
 
 ## [2.2.0]
 ### Added
@@ -10,7 +12,6 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 - Small fixes for contributing guide
-- Use imperative mood in changelog entry
 
 ### Dependencies
 - Bumps `typescript` from 4.9.3 to 4.9.4