+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [UNRELEASED]
+
+## [2.2.1]
 ### Fixed
 - Fixes an issue where a duplicate entry was added when a previous entry to update was found.